requirements.txt
# Tłumacz PDF/TXT → PL — zależności
# Instalacja: pip install -r requirements.txt
# PyTorch z CUDA: pip install torch --index-url https://download.pytorch.org/whl/cu121

torch>=2.1.0
ctranslate2>=4.0.0
transformers>=4.38.0
sentencepiece>=0.1.99
sacremoses>=0.0.53

pymupdf>=1.23.0
pdfplumber>=0.10.0

python-docx>=1.1.0
Pillow>=10.0.0

customtkinter>=5.2.2
